Player can generate ticket
On the player dashboard they will have a option to generate ticket

//TODO work
Now the ticket generated by the user we have to save it to the databaseSo that that ticket got saved throughout the game and whenever loaded that gaeme the same ticket displays and the same ticket can be viewed by the host of the game.

Further we had to make sure that the player can modify the number of tickets they have created.

Designing of the tickets as well as the ticket list with the host viewing table for all the players ticket deisging is left

Last work the tickets also got marked with the numbers that are called