Styling done
Styling is improved and all the bug has been improved ticket designing has been updated top row is also styled properly and the loading for the game is also shown properly even when the game is not loaded for a certain period of time.
Reload Button has also been provided for the player to fetch the ticket until its not given.

ToDo:

Save the marked numbers by the player in the session storage

await for the dividend verify call resposne and revert accordingly
for e.g if its success then show success else show boogie.